automatically computing output filename
Output filenames are automatically computed
from the MaxQuant input filename that can be
found in column "Raw.file" of "allPeptides.txt".
If this column has a single filename, e.g.
"AGA150" then the output filename is named
"AGA150.mzTab".
If the MaxQuant results are based on multiple
input files (label free) then the ouput filename is named:
"AGA150_etal.mzTab", where "AGA150" is the entry
in the first row of column "Raw.file" in "allPeptides.txt"
as written by MaxQuant.