added command line parsing with error handling
Folder name containing max quant results is
parsed as first argument. Calling the R script
with zero or >= 2 arguments raises an error
detailing proper usage.
After parsing the folder name, existance
is verified using `dir.exists` and if we can
not find a folder at this path
an error about that is raised also.
We might tighten this a little more by also checking
for existance of all relevant files inside
the folder.